PROMS dictionary

Data dictionary of Patient Reported Outcomes and Measures (PROMS) self-report survey data. Subject to change.

Usage

dictionary_proms

Format

A data frame with 8 variables:

instrument

Survey instrument

instrument_description

Survey instrument description

variable_type

Type of survey question (one of multiple choice, select all that apply, write-in, etc.)

variable

Variable name that references instrument and question

label

Variable label with question text

values_original

Reponse values as original programmed

values_clean

Response values cleaned with lower case and extra spaces removed

col_type

How the variable is stored in R

Examples

head(dictionary_proms)
#> # A tibble: 6 × 8
#>   instrument instrument_description variable_type variable label values_original
#>   <chr>      <chr>                  <chr>         <chr>    <chr> <chr>          
#> 1 Not Appli… Not Applicable         internal      study_id Subj… NA             
#> 2 Not Appli… Not Applicable         internal      site_co… Char… NA             
#> 3 Not Appli… Not Applicable         internal      site_na… Site… NA             
#> 4 Not Appli… Not Applicable         internal      site_na… Full… NA             
#> 5 Not Appli… Not Applicable         internal      site_na… Shor… NA             
#> 6 Not Appli… Not Applicable         internal      site_co… Coun… NA             
#> # ℹ 2 more variables: values_clean <chr>, col_type <chr>