Treatment mapping

A dataset containing treatment mapping specific to PCCTC studies, intended to be updated regularly. If a mapping needs to be corrected or added, please file an issue or contact PCCTC. Note that this mapping should already be in the ca_cm_derived table.

Usage

treatment_map

Format

A data frame with 5 variables:

treatment

Treatment as submitted to the study

treatment_lower

Treatment as submitted to the study lower case

treatment_recode

Assigned treatment standardized coding

treatment_category

Assigned treatment category

is_taxel

Indicates if treatment is a taxel (one of Cabazitaxel, Docetaxel, Paclitaxel, Tesetaxel)

is_nha

Indicates if treatment is a novel hormone agent (one of Abiraterone, Enzalutamide, Darolutamide, Apalutamide)

Examples

print(treatment_map)
#> # A tibble: 169 × 6
#>    treatment treatment_lower treatment_recode treatment_category is_taxel is_nha
#>    <chr>     <chr>           <chr>            <chr>              <lgl>    <lgl> 
#>  1 Casodex … casodex (bical… Bicalutamide     ADT                FALSE    FALSE 
#>  2 Nilandro… nilandron (nil… Nilutamide       ADT                FALSE    FALSE 
#>  3 Flutamide flutamide       Flutamide        ADT                FALSE    FALSE 
#>  4 Degarelix degarelix       Degarelix        ADT                FALSE    FALSE 
#>  5 Bicaluta… bicalutamide    Bicalutamide     ADT                FALSE    FALSE 
#>  6 Lupron    lupron          Leuprolide       ADT                FALSE    FALSE 
#>  7 Testoste… testosterone    Testosterone     Steroid            FALSE    FALSE 
#>  8 Zoladex … zoladex (goser… Goserelin        ADT                FALSE    FALSE 
#>  9 Leuproli… leuprolide      Leuprolide       ADT                FALSE    FALSE 
#> 10 Zometa    zometa          Zoledronic acid  Bisphosphonates    FALSE    FALSE 
#> # ℹ 159 more rows